+++ NEWS.md
@@ -1,3 +1,6 @@
+## gDRutils 1.5.13 - 2025-03-12
+* improve logic in `get_assay_req_uniq_cols`
+
 ## gDRutils 1.5.12 - 2025-02-21
 * refactor `average_biological_replicates`